refactor: restructures hooks
