+++ src/components/ProfileModal/EmailModal.tsx
@@ -1,15 +1,16 @@
 import { FC } from 'react'
-import { Logo } from '../Logo/Logo'
-import { Button } from '../Button/Button'
 import { SubmitHandler, useForm } from 'react-hook-form'
-
-import styles from './style.module.css'
 import { useNavigate } from 'react-router-dom'
-import { useAppSelector } from '../../hooks/appHooks'
+
+import { useChangeEmailMutation } from '../../api/auth.api'
+import { useAppCookies, useAppSelector } from '../../hooks/appHooks'
 import { ROUTES } from '../../routes'
 import { selectCurrentUser } from '../../slices/currentUserSlice'
-import { useChangeEmailMutation } from '../../api/auth.api'
-import { useAppCookies } from '../../hooks/userHooks'
+import { AppCookies } from '../../types'
+import { Button } from '../Button/Button'
+import { Logo } from '../Logo/Logo'
+
+import styles from './style.module.css'
 
 type EmailModalProps = {
   setIsOpened: Function
@@ -40,14 +41,12 @@ export const EmailModal: FC<EmailModalProps> = ({ setIsOpened }) => {
 
     try {
       const res = await changeEmail({ idToken: user.idToken, email: data.email }).unwrap()
-      setCookies({ ...res })
+      setCookies({ ...res } as AppCookies)
       setIsOpened(false)
     } catch (error) {
       console.error('Change email failed -->', error)
       navigate(ROUTES.login)
     }
-
-    console.log(data)
   }
 
   return (
