+++ src/components/ProfileModal/PasswordModal.tsx
@@ -4,10 +4,10 @@ import { SubmitHandler, useForm } from 'react-hook-form'
 import { useNavigate } from 'react-router-dom'
 
 import { useChangePasswordMutation } from '../../api/auth.api'
-import { useAppSelector } from '../../hooks/appHooks'
-import { useAppCookies } from '../../hooks/userHooks'
+import { useAppCookies, useAppSelector } from '../../hooks/appHooks'
 import { ROUTES } from '../../routes'
 import { selectCurrentUser } from '../../slices/currentUserSlice'
+import { AppCookies } from '../../types'
 import { Button } from '../Button/Button'
 import { Logo } from '../Logo/Logo'
 
@@ -47,7 +47,7 @@ export const PasswordModal: FC<PasswordModalProps> = ({ setIsOpened }) => {
       const res = await changePassword({
         idToken: user.idToken, password: data.password
       }).unwrap()
-      setCookies({ ...res })
+      setCookies({ ...res } as AppCookies)
       setIsOpened(false)
     } catch (error) {
       console.error('Change password failed -->', error)
