+++ src/components/User/User.tsx
@@ -2,7 +2,8 @@ import classNames from 'classnames'
 import { FC, useState } from 'react'
 import { useDispatch } from 'react-redux'
 import { useNavigate } from 'react-router-dom'
-import { useAppCookies } from '../../hooks/userHooks'
+
+import { useAppCookies } from '../../hooks/appHooks'
 import { ROUTES } from '../../routes'
 import { deleteCurrentUser } from '../../slices/currentUserSlice'
 import { FirebaseUserRESTAPI } from '../../types'
