+++ src/hooks/userHooks.ts
@@ -1,136 +1,12 @@
-import { useEffect, useState } from 'react'
-import {
-  createUserWithEmailAndPassword,
-  signInWithEmailAndPassword,
-  signOut,
-  updateEmail as firebaseUpdateEmail,
-  updatePassword as firebaseUpdatePassword
-} from 'firebase/auth'
-import { get, ref } from 'firebase/database'
 import { merge } from 'lodash'
+import { useEffect, useState } from 'react'
 
-import auth from "../db/auth"
-import db from "../db/db"
 import { useGetCourseQuery, useGetCoursesQuery } from '../api/courses.api'
 import { useGetUserCourseQuery, useGetUserCoursesQuery } from '../api/users.api'
 import { CourseData } from '../types'
-import { useDispatch } from 'react-redux'
-import { initialState, setUser } from '../slices/userSlice'
-import { useCookies } from 'react-cookie'
-import { useAppDispatch } from './appHooks'
-import { updateCurrentUser } from '../slices/currentUserSlice'
 
-export const useAuth = () => {
-  const noop = () => {}  // заглушка для колбэков
-  const dispatch = useDispatch()
-
-  // залогиниться
-  const signIn = (
-    username: string,
-    password: string,
-    successCallback = noop,
-    errorCallback = noop
-    ) => {
-    dispatch(setUser({ ...initialState }))
-    signInWithEmailAndPassword(auth, username, password)
-      .then((userCredential) => {
-        const user = userCredential.user;
-        console.log('User signed in: ', user)
-        dispatch(setUser({
-          uid: user.uid,
-          email: user.email,
-          displayName: user.displayName,
-          isLoading: false
-        }))
-        successCallback() // на всякий случай
-      })
-      .catch((error) => {
-        console.error(
-          `sign-in failed: error.code=${error.code}, error.message=${error.message}`
-        )
-        errorCallback() // на всякий случай
-      })   
-  }
-
-  // разлогиниться
-  const logOut = (successCallback = noop, errorCallback = noop) => {
-    signOut(auth).then(() => {
-      console.log('Sign-out successful!')
-      successCallback() // на всякий случай
-    }).catch((error) => {
-      console.error('Sign-out failed!')
-      errorCallback() // на всякий случай
-    })
-  }
-
-  // зарегистировать нового пользователя
-  const signUp = (
-    username: string,
-    password: string,
-    successCallback = noop,
-    errorCallback = noop
-    ) => {
-    createUserWithEmailAndPassword(auth, username, password)
-      .then((userCredential) => {
-        // Signed in 
-        const user = userCredential.user;
-        console.log('User created -->', user)
-        successCallback() // на всякий случай
-      })
-      .catch((error) => {
-        console.error(
-          `Error creating a user: error.code=${error.code}, error.message=${error.message}`
-        )
-        errorCallback() // на всякий случай
-      })
-  }
-
-  return { signIn, logOut, signUp }
-}
-
-export const useManageUser = () => {
-  const noop = () => {}  // заглушка для колбэков
-
-  const updateEmail = (
-    newEmail: string,
-    successCallback = noop,
-    errorCallback = noop
-  ) => {
-    if (auth.currentUser) {
-      firebaseUpdateEmail(auth.currentUser, newEmail).then(() => {
-        console.log('Email updated')
-        successCallback()
-      }).catch((error) => {
-        console.error(
-          `update email failed: error.code=${error.code}, error.message=${error.message}`
-        )
-        errorCallback()
-      })
-    }
-  }
-
-  const updatePassword = (
-    newPassword: string,
-    successCallback = noop,
-    errorCallback = noop
-  ) => {
-    if (auth.currentUser) {
-      firebaseUpdatePassword(auth.currentUser, newPassword).then(() => {
-        console.log('Password updated')
-        successCallback()
-      }).catch((error) => {
-        console.error(
-          `update password failed: error.code=${error.code}, error.message=${error.message}`
-        )
-        errorCallback()
-      })
-  
-    }
-  }
-  
-  return { updateEmail, updatePassword }
-}
-
+// возвращает список ключей, не равных null
+// это необходимо для очистки сырой информации из БД
 const getValidKeys = (obj: object) => {
   const validKeys = []
   for (let [curKey, curValue] of Object.entries(obj)) {
@@ -139,6 +15,7 @@ const getValidKeys = (obj: object) => {
   return validKeys
 }
 
+// возвращает курсы заданного пользователя (без данных из /users)
 export const useUserCourses = (uid: string) => {
   const { data: courses, isLoading: isCoursesLoading } = useGetCoursesQuery()
   const { data: userCoursesData, isLoading: isUserCoursesLoading } = useGetUserCoursesQuery(uid)
@@ -165,6 +42,7 @@ export const useUserCourses = (uid: string) => {
   return { data: userCourses, isLoading }
 }
 
+// возвращает курсы с доп полем subscription чтобы добавлять/удалять курсы для пользователяя
 export const useCoursesWithSubscription = (uid: string) => {
   const { data: courses, isLoading: isCoursesLoading } = useGetCoursesQuery()
   const { data: userCoursesData, isLoading: isUserCoursesLoading } = useGetUserCoursesQuery(uid)
@@ -199,6 +77,7 @@ export const useCoursesWithSubscription = (uid: string) => {
   return { data: coursesWithSubscription, isLoading }
 }
 
+// полнные даннные по заданному курсу пользователя
 export const useUserCourse = (uid: string | null, courseId: number) => {
   const { data: course } = useGetCourseQuery(courseId)
   const { 
@@ -230,77 +109,4 @@ export const useUserCourse = (uid: string | null, courseId: number) => {
   }, [userCourseData, course, isUserCourseLoading, uid])
   
   return { data: userCourse, isLoading: isUserCourseLoading, error, isError }
-}
-
-
-// TODO убрать, если не используется
-export const useUserWorkoutStatus = (uid: string, courseId: number, workoutId: number) => {
-  const [status, setStatus] = useState()
-
-  useEffect(() => {
-
-  }, )
-  
-  const workoutRef = ref(db, `users/${uid}/courses/${courseId}/workouts/${workoutId}/done`)
-  get(workoutRef).then((snapshot) => {
-    if (snapshot.exists()) {
-      setStatus(snapshot.val())
-    } else {
-      console.warn("No data available")
-    }
-  }).catch((error) => {
-    console.error(error)
-  })
-
-  return status
-}
-
-export const useLoadCredentialsFromCookies = () => {
-  const { cookies } = useAppCookies()
-  const dispatch = useAppDispatch()
-  const { idToken,  refreshToken, localId, email } = cookies
-
-  const loadCredentials = () => {
-    if (cookies && idToken) {
-      console.log('updating currentUser')
-      dispatch(updateCurrentUser({
-        idToken,
-        refreshToken,
-        localId,
-        email
-      }))
-    } else {
-      console.warn('no credentials found in cookies');
-    } 
-  }
-  
-  return { loadCredentials }
-}
-
-type AppCookies = {
-  idToken?: string
-  refreshToken?: string
-  localId?: string
-  email?: string
-}
-
-export const useAppCookies = () => {
-  const [ cookies, setCookie, removeCookie ] = useCookies(['idToken', 'refreshToken', 'localId', 'email'])
-
-  const setCookies = (cookiesArg: AppCookies) => {
-    const { idToken,  refreshToken, localId, email } = cookiesArg
-    setCookie('idToken', idToken)
-    setCookie('refreshToken', refreshToken)
-    setCookie('localId', localId)
-    setCookie('email', email)
-    }
-
-  const removeCookies = () => {
-    removeCookie('idToken')
-    removeCookie('refreshToken')
-    removeCookie('localId')
-    removeCookie('email')
-  }
-
-  return { cookies, setCookies, removeCookies }
 }
